requirements.txt
# General modules
django
python-decouple
Pillow
djangorestframework

# Third party modules
markdown 
django-filter
drf-yasg[validation]
setuptools
djangorestframework-simplejwt

# Email third party modules
django-mail-templated

# refactoring and tests
pytest
pytest-django